+++ lib/redmine.rb
@@ -293,10 +293,10 @@ Redmine::MenuManager.map :project_menu do |menu|
                       :if => Proc.new { |p| p.shared_versions.any? },
                       :html => {:class => "icon2 icon-new-planning-element"}
 
-  menu.push :work_packages, { :controller => '/work_packages', :action => 'index' },
-                            :param => :project_id,
-                            :caption => :label_work_package_plural,
-                            :html => {:class => "icon2 icon-copy"}
+  menu.push :work_packages, { controller: '/work_packages', action: 'index', set_filter: 1 },
+                            param: :project_id,
+                            caption: :label_work_package_plural,
+                            html: {class: "icon2 icon-copy"}
 
   menu.push :new_work_package, { :controller => '/work_packages', :action => 'new'},
                                :param => :project_id,